requirements_updated.txt
Flask==3.0.3
opencv-python==4.13.0.92
PyMuPDF==1.27.2.2
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.1.0
reportlab==4.0.9
pandas>=2.0
SQLAlchemy>=2.0
openpyxl>=3.0
easyocr>=1.7.0
torch>=2.0.0
transformers>=4.30.0
spacy>=3.5.0
